+++ rotates.c
@@ -6,7 +6,7 @@
 /*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 12:44:47 by adantas-          #+#    #+#             */
-/*   Updated: 2023/02/13 13:25:02 by adantas-         ###   ########.fr       */
+/*   Updated: 2023/02/14 15:28:49 by adantas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	rot_a(t_stack *stks)
 {
-	int	i;
-	int	tmp;
+	size_t	i;
+	int		tmp;
 
 	if (stks->sz_a < 2)
 		return ;
@@ -29,8 +29,8 @@ void	rot_a(t_stack *stks)
 
 void	rot_b(t_stack *stks)
 {
-	int	i;
-	int	tmp;
+	size_t	i;
+	int		tmp;
 
 	if (stks->sz_b < 2)
 		return ;
@@ -44,8 +44,8 @@ void	rot_b(t_stack *stks)
 
 void	rot_ab(t_stack *stks)
 {
-	int	i;
-	int	tmp;
+	size_t	i;
+	int		tmp;
 
 	if (stks->sz_a < 2 || stks->sz_b < 2)
 		return ;
